% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_TSO.R
\name{display_TSO}
\alias{display_TSO}
\title{Display tif, segmentation, and overlay images in a row}
\usage{
display_TSO(ls, directory)
}
\arguments{
\item{ls}{A list of three file names.}

\item{directory}{A directory in which the files with the names listed in the
ls paramter exist.}
}
\description{
This function takes in a list of three file names: the name of a tif file,
a segmentation png file, and an overlay png file. It will output a visual
of the three files side-by-side.
}
