% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN_annotate.R
\name{KNN_annotate}
\alias{KNN_annotate}
\title{Use binary k-nearest-neighbors classification to annotate cells}
\usage{
KNN_annotate(df, chan_oi, k, features)
}
\arguments{
\item{df}{A "Results" data frame saved from ImageJ.}

\item{chan_oi}{The name of the channel of interest.}

\item{k}{The number of neighbors used for classification.}

\item{features}{A vector of features from the data frame to use for
classification. These will be column names.}
}
\description{
This function will use a training set in the current directory that
corresponds to a given data frame and use it to annotate all of the cells in
that data frame. This function is tailored for use with this package's
workflow. It is not recommended for use with other types of data.
}
