% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_hist_explore.R
\name{stat_hist_explore}
\alias{stat_hist_explore}
\title{Get an exploratory histogram for a given column in a data frame.}
\usage{
stat_hist_explore(df, channel, column)
}
\arguments{
\item{df}{A "Results" data frame saved from ImageJ.}

\item{channel}{A current channel of interest that is being examined in the
images.}

\item{column}{The column in the data frame that is populated with data of a
certain statistic.}
}
\description{
This function will take a data frame and an indicated column and will output
an exploratory histogram for the data in that column, splitting by if cells
are "channel positive" or are "normal" cells. The channel must be indicated
as well.
}
